Fixed memory leaks
Switched to raw pointers for components to hold a reference
Did this because I want a bi-directional relationship between gameobjects and components that allows me to modify them in any direction.
Research told me this is the only way to do it unless you work with an event system, of which I have no knowledge yet.
Could look into it for the final product!